requirements.txt
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.2
docutils==0.18.1
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Kivy
outcome==1.1.0
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
pypiwin32==223
PySocks==1.7.1
pytest==7.1.2
pywin32==303
PyYAML==6.0
requests==2.27.1
rich==12.0.1
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
websockets==10.2
wsproto==1.1.0
